requirements.txt
s2t2-game-utils
python-dotenv